+++ modules/Grades/InputFinalGrades.php
@@ -2,6 +2,7 @@
 
 require_once 'modules/Grades/includes/ClassRank.inc.php';
 require_once 'modules/Grades/includes/Grades.fnc.php';
+require_once 'modules/Grades/includes/FinalGrades.inc.php';
 require_once 'ProgramFunctions/TipMessage.fnc.php';
 
 if ( ! empty( $_REQUEST['period'] ) )
@@ -227,248 +228,21 @@ if ( $_REQUEST['modfunc'] === 'gradebook' )
 {
 	if ( ! empty( $_REQUEST['mp'] ) )
 	{
-		$gradebook_config = ProgramUserConfig( 'Gradebook' );
-
-		$_ROSARIO['_makeLetterGrade']['courses'][$course_period_id] = DBGet( "SELECT DOES_BREAKOFF,GRADE_SCALE_ID
-			FROM course_periods
-			WHERE COURSE_PERIOD_ID='" . (int) $course_period_id . "'" );
-
-		require_once 'ProgramFunctions/_makeLetterGrade.fnc.php';
-
-		if ( GetMP( $_REQUEST['mp'], 'MP' ) == 'QTR' || GetMP( $_REQUEST['mp'], 'MP' ) == 'PRO' )
+		if ( in_array( GetMP( $_REQUEST['mp'], 'MP' ), [ 'QTR', 'PRO' ] ) )
 		{
-			// Note: The 'active assignment' determination is not fully correct.  It would be easy to be fully correct here but the same determination
-			// as in Grades.php is used to avoid apparent inconsistencies in the grade calculations.  See also the note at top of Grades.php.
-			$extra['SELECT_ONLY'] = "s.STUDENT_ID, gt.ASSIGNMENT_TYPE_ID,sum(" .
-			db_case( [ 'gg.POINTS', "'-1'", "'0'", 'gg.POINTS' ] ) . ") AS PARTIAL_POINTS,sum(" .
-			db_case( [ 'gg.POINTS', "'-1'", "'0'", 'ga.POINTS' ] ) . ") AS PARTIAL_TOTAL,gt.FINAL_GRADE_PERCENT";
-
-			if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] ) )
-			{
-				// @since 11.0 Add Weight Assignments option
-				$extra['SELECT_ONLY'] .= ",sum(" . db_case( [ 'ga.WEIGHT', "''", "'0'", "ga.WEIGHT" ] ) . ") AS PARTIAL_WEIGHT,
-					sum((gg.POINTS/ga.POINTS)*ga.WEIGHT) AS PARTIAL_WEIGHTED_GRADE";
-			}
-
-			$extra['FROM'] = " JOIN gradebook_assignments ga ON
-				(((ga.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID
-						OR ga.COURSE_ID=cp.COURSE_ID)
-						AND ga.STAFF_ID=cp.TEACHER_ID)
-					AND ga.MARKING_PERIOD_ID='" . UserMP() . "')
-				LEFT OUTER JOIN gradebook_grades gg ON
-				(gg.STUDENT_ID=s.STUDENT_ID
-					AND gg.ASSIGNMENT_ID=ga.ASSIGNMENT_ID
-					AND gg.COURSE_PERIOD_ID=cp.COURSE_PERIOD_ID),gradebook_assignment_types gt";
-
-			// Check Current date.
-			$extra['WHERE'] = " AND gt.ASSIGNMENT_TYPE_ID=ga.ASSIGNMENT_TYPE_ID
-				AND gt.COURSE_ID=cp.COURSE_ID
-				AND (gg.POINTS IS NOT NULL
-					OR (ga.ASSIGNED_DATE IS NULL OR CURRENT_DATE>=ga.ASSIGNED_DATE)
-					AND (ga.DUE_DATE IS NULL OR CURRENT_DATE>=ga.DUE_DATE)
-					OR CURRENT_DATE>(SELECT END_DATE
-						FROM school_marking_periods
-						WHERE MARKING_PERIOD_ID=ga.MARKING_PERIOD_ID))";
-
-			// Check Student enrollment.
-			$extra['WHERE'] .= " AND (gg.POINTS IS NOT NULL
-				OR ga.DUE_DATE IS NULL
-				OR ((ga.DUE_DATE>=ss.START_DATE
-					AND (ss.END_DATE IS NULL OR ga.DUE_DATE<=ss.END_DATE))
-				AND (ga.DUE_DATE>=ssm.START_DATE
-					AND (ssm.END_DATE IS NULL OR ga.DUE_DATE<=ssm.END_DATE))))";
-
-			if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] ) )
-			{
-				// @since 11.0 Add Weight Assignments option
-				// Exclude Extra Credit assignments.
-				$extra['WHERE'] .= " AND ga.POINTS>0";
-			}
-
-			if ( GetMP( $_REQUEST['mp'], 'MP' ) === 'PRO' )
-			{
-				// FJ: limit Assignments to the ones due during the Progress Period.
-				$extra['WHERE'] .= " AND ((ga.ASSIGNED_DATE IS NULL OR (SELECT END_DATE
-					FROM school_marking_periods
-					WHERE MARKING_PERIOD_ID='" . (int) $_REQUEST['mp'] . "')>=ga.ASSIGNED_DATE)
-					AND (ga.DUE_DATE IS NULL
-						OR (SELECT END_DATE
-							FROM school_marking_periods
-							WHERE MARKING_PERIOD_ID='" . (int) $_REQUEST['mp'] . "')>=ga.DUE_DATE
-						AND (SELECT START_DATE
-							FROM school_marking_periods
-							WHERE MARKING_PERIOD_ID='" . (int) $_REQUEST['mp'] . "')<=ga.DUE_DATE))";
-			}
-
-			$extra['GROUP'] = "gt.ASSIGNMENT_TYPE_ID,gt.FINAL_GRADE_PERCENT,s.STUDENT_ID";
-
-			$extra['group'] = [ 'STUDENT_ID' ];
-
-			$points_RET = GetStuList( $extra );
-			//echo '<pre>'; var_dump($points_RET); echo '</pre>';
-
-			unset( $extra );
-
-			if ( ! empty( $points_RET ) )
-			{
-				foreach ( (array) $points_RET as $student_id => $student )
-				{
-					$total = $total_percent = $total_weighted_grade = $total_weights = 0;
-
-					foreach ( (array) $student as $partial_points )
-					{
-						/**
-						 * Do not include Extra Credit assignments
-						 * when Total Points is 0 for the Type
-						 * if the Gradebook is configured to Weight Grades:
-						 * Division by zero is impossible.
-						 */
-						if ( $partial_points['PARTIAL_TOTAL'] != 0
-							|| empty( $gradebook_config['WEIGHT'] ) )
-						{
-							$total += $partial_points['PARTIAL_POINTS'] * ( ! empty( $gradebook_config['WEIGHT'] ) ?
-								$partial_points['FINAL_GRADE_PERCENT'] / $partial_points['PARTIAL_TOTAL'] :
-								1
-							);
-
-							$total_percent += ( ! empty( $gradebook_config['WEIGHT'] ) ?
-								$partial_points['FINAL_GRADE_PERCENT'] :
-								$partial_points['PARTIAL_TOTAL']
-							);
-
-							if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] ) )
-							{
-								// @since 11.0 Add Weight Assignments option
-								$total_weighted_grade += ( ! empty( $gradebook_config['WEIGHT'] ) ?
-									$partial_points['FINAL_GRADE_PERCENT'] * $partial_points['PARTIAL_WEIGHTED_GRADE'] :
-									$partial_points['PARTIAL_WEIGHTED_GRADE'] );
-
-								$total_weights += ( ! empty( $gradebook_config['WEIGHT'] ) ?
-									$partial_points['FINAL_GRADE_PERCENT'] * $partial_points['PARTIAL_WEIGHT'] :
-									$partial_points['PARTIAL_WEIGHT'] );
-							}
-						}
-					}
-
-					if ( $total_percent != 0 )
-					{
-						$total /= $total_percent;
-					}
-
-					if ( ! empty( $gradebook_config['WEIGHT_ASSIGNMENTS'] )
-						&& $total_weights > 0 )
-					{
-						// @since 11.0 Add Weight Assignments option
-						$total = $total_weighted_grade / $total_weights;
-					}
-
-					$import_RET[$student_id] = [
-						1 => [
-							'REPORT_CARD_GRADE_ID' => _makeLetterGrade( $total, $course_period_id, 0, 'ID' ),
-							'GRADE_PERCENT' => round( 100 * $total, 1 ),
-						],
-					];
-				}
-			}
+			$import_RET = FinalGradesQtrOrProCalculate(
+				$course_period_id,
+				$_REQUEST['mp']
+			);
 		}
-		elseif ( GetMP( $_REQUEST['mp'], 'MP' ) == 'SEM' || GetMP( $_REQUEST['mp'], 'MP' ) == 'FY' )
+		elseif ( in_array( GetMP( $_REQUEST['mp'], 'MP' ), [ 'SEM', 'FY' ] ) )
 		{
-			if ( GetMP( $_REQUEST['mp'], 'MP' ) == 'SEM' )
-			{
-				$mp_RET = DBGet( "SELECT MARKING_PERIOD_ID,'Y' AS DOES_GRADES
-				FROM school_marking_periods
-				WHERE MP='QTR'
-				AND PARENT_ID='" . (int) $_REQUEST['mp'] . "'
-				UNION
-				SELECT MARKING_PERIOD_ID,NULL AS DOES_GRADES
-				FROM school_marking_periods
-				WHERE MP='SEM'
-				AND MARKING_PERIOD_ID='" . (int) $_REQUEST['mp'] . "'" );
-				$prefix = 'SEM-';
-			}
-			else
-			{
-				$mp_RET = DBGet( "SELECT q.MARKING_PERIOD_ID,'Y' AS DOES_GRADES
-				FROM school_marking_periods q,school_marking_periods s
-				WHERE q.MP='QTR'
-				AND s.MP='SEM'
-				AND q.PARENT_ID=s.MARKING_PERIOD_ID
-				AND s.PARENT_ID='" . (int) $_REQUEST['mp'] . "'
-				UNION
-				SELECT MARKING_PERIOD_ID,DOES_GRADES
-				FROM school_marking_periods
-				WHERE MP='SEM'
-				AND PARENT_ID='" . (int) $_REQUEST['mp'] . "'
-				UNION
-				SELECT MARKING_PERIOD_ID,NULL AS DOES_GRADES
-				FROM school_marking_periods
-				WHERE MP='FY'
-				AND MARKING_PERIOD_ID='" . (int) $_REQUEST['mp'] . "'" );
-				$prefix = 'FY-';
-			}
-
-			$mps = '';
-
-			foreach ( (array) $mp_RET as $mp )
-			{
-				if ( $mp['DOES_GRADES'] === 'Y' )
-				{
-					$mps .= "'" . $mp['MARKING_PERIOD_ID'] . "',";
-				}
-			}
-
-			$mps = mb_substr( $mps, 0, -1 );
-
-			$percents_RET = DBGet( "SELECT STUDENT_ID,GRADE_PERCENT,MARKING_PERIOD_ID
-				FROM student_report_card_grades
-				WHERE COURSE_PERIOD_ID='" . (int) $course_period_id . "'
-				AND MARKING_PERIOD_ID IN (" . $mps . ")", [], [ 'STUDENT_ID' ] );
-
-			foreach ( (array) $percents_RET as $student_id => $percents )
-			{
-				$total = $total_percent = 0;
-
-				foreach ( (array) $percents as $percent )
-				{
-					if ( ! isset( $gradebook_config[$prefix . $percent['MARKING_PERIOD_ID']] ) )
-					{
-						// @since 11.5 Add "Final Grading Percentages are not configured." warning
-						$warning['config_percent'] = _( 'Final Grading Percentages are not configured.' );
-
-						if ( AllowUse( 'Grades/Configuration.php' ) )
-						{
-							$warning['config_percent'] .= ' <a href="Modules.php?modname=Grades/Configuration.php">' .
-								_( 'Configuration' ) . '</a>';
-						}
-					}
-
-					$total += $percent['GRADE_PERCENT'] *
-						issetVal( $gradebook_config[$prefix . $percent['MARKING_PERIOD_ID']] );
-
-					$total_percent += $gradebook_config[$prefix . $percent['MARKING_PERIOD_ID']];
-				}
-
-				if ( $total_percent != 0 )
-				{
-					$total /= $total_percent;
-				}
-
-				$import_RET[$student_id] = [
-					1 => [
-						'REPORT_CARD_GRADE_ID' => _makeLetterGrade( $total / 100, $course_period_id, 0, 'ID' ),
-						'GRADE_PERCENT' => round( $total, 1 ),
-					],
-				];
-
-				// FJ automatic comment on yearly grades.
-
-				if ( GetMP( $_REQUEST['mp'], 'MP' ) === 'FY' )
-				{
-					// FJ use Report Card Grades comments.
-					$comment = _makeLetterGrade( $total / 100, $course_period_id, 0, 'COMMENT' );
-					$import_comments_RET[$student_id][1]['COMMENT'] = $comment;
-				}
-			}
+			// Do not fail on warning "Final Grading Percentages are not configured."
+			$import_RET = FinalGradesSemOrFYCalculate(
+				$course_period_id,
+				$_REQUEST['mp'],
+				'continue'
+			);
 		}
 	}
 
@@ -621,18 +395,16 @@ if ( ! empty( $_REQUEST['values'] )
 					$letter = $grades_RET[$grade][1]['TITLE'];
 					$weighted = $grades_RET[$grade][1]['WEIGHTED_GP'];
 					$unweighted = $grades_RET[$grade][1]['UNWEIGHTED_GP'];
+					$scale = $grades_RET[$grade][1]['GP_SCALE'];
+					$gp_passing = $grades_RET[$grade][1]['GP_PASSING_VALUE'];
 
 					// FJ add precision to year weighted GPA if not year course period.
 
 					if ( GetMP( $_REQUEST['mp'], 'MP' ) === 'FY'
 						&& $course_period_mp !== 'FY' )
 					{
-						$weighted = $percent / 100 * $grades_RET[$grade][1]['GP_SCALE'];
+						$weighted = $percent / 100 * $scale;
 					}
-
-					$scale = $grades_RET[$grade][1]['GP_SCALE'];
-
-					$gp_passing = $grades_RET[$grade][1]['GP_PASSING_VALUE'];
 				}
 				else
 				{
@@ -659,7 +431,7 @@ if ( ! empty( $_REQUEST['values'] )
 				$update_columns = [
 					'GRADE_PERCENT' => $percent,
 					'REPORT_CARD_GRADE_ID' => $grade,
-					'GRADE_LETTER' => $letter,
+					'GRADE_LETTER' => DBEscapeString( $letter ),
 					'WEIGHTED_GP' => $weighted,
 					'UNWEIGHTED_GP' => $unweighted,
 					'GP_SCALE' => $scale,
@@ -674,25 +446,23 @@ if ( ! empty( $_REQUEST['values'] )
 				$grade = $columns['grade'];
 				$letter = $grades_RET[$grade][1]['TITLE'];
 				$weighted = $grades_RET[$grade][1]['WEIGHTED_GP'];
+				$unweighted = $grades_RET[$grade][1]['UNWEIGHTED_GP'];
+				$scale = $grades_RET[$grade][1]['GP_SCALE'];
+				$gp_passing = $grades_RET[$grade][1]['GP_PASSING_VALUE'];
 
 				// FJ add precision to year weighted GPA if not year course period.
 
 				if ( GetMP( $_REQUEST['mp'], 'MP' ) === 'FY'
 					&& $course_period_mp !== 'FY' )
 				{
-					$weighted = $percent / 100 * $grades_RET[$grade][1]['GP_SCALE'];
+					$weighted = $percent / 100 * $scale;
 				}
 
-				$unweighted = $grades_RET[$grade][1]['UNWEIGHTED_GP'];
-
-				$scale = $grades_RET[$grade][1]['GP_SCALE'];
-
-				$gp_passing = $grades_RET[$grade][1]['GP_PASSING_VALUE'];
 
 				$update_columns = [
 					'GRADE_PERCENT' => $percent,
 					'REPORT_CARD_GRADE_ID' => $grade,
-					'GRADE_LETTER' => $letter,
+					'GRADE_LETTER' => DBEscapeString( $letter ),
 					'WEIGHTED_GP' => $weighted,
 					'UNWEIGHTED_GP' => $unweighted,
 					'GP_SCALE' => $scale,